Format HCL on output.
Also adds a flag to use TF 0.12 formatter.
The TF 0.12 formatter is less aggressive than earlier versions, so default to using TF 0.11 formatter for now until TF 0.12 goes GA.

This change also moves log output setup into it’s own go file.
