+++ main.go
@@ -2,14 +2,10 @@ package main
 
 import (
 	"fmt"
-	"os"
-	"strings"
-
+	"github.com/hashicorp/hcl/hcl/printer"
+	"github.com/hashicorp/hcl2/hclwrite"
 	flag "github.com/spf13/pflag"
 
-	"github.com/hashicorp/hcl2/hclwrite"
-
-	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 )
 
@@ -28,31 +24,21 @@ var (
 	includeUnsupported bool
 	noColor            bool
 	overwriteExisting  bool
+	tf12format bool
 )
 
 func init() {
 	// init command line flags
 	flag.BoolVarP(&overwriteExisting, "overwrite-existing", "x", false, "allow overwriting existing output file(s)")
 	flag.BoolVarP(&debug, "debug", "d", false, "enable debug output")
-	flag.StringVarP(&input, "filepath", "f", "-", `file or directory that contains the YAML configuration to convert. Use "-" to read from stdin.`)
+	flag.StringVarP(&input, "filepath", "f", "-", `file or directory that contains the YAML configuration to convert. Use "-" to read from stdin`)
 	flag.StringVarP(&output, "output", "o", "-", `file or directory where Terraform config will be written`)
 	flag.BoolVarP(&includeUnsupported, "include-unsupported", "I", false, `set to true to include unsupported Attributes / Blocks in the generated TF config`)
+	flag.BoolVarP(&tf12format, "tf12format", "F", false, `Use Terraform 0.12 formatter`)
 
 	flag.Parse()
 
-	// Setup Console Output
-	output := zerolog.ConsoleWriter{Out: os.Stderr}
-	output.FormatLevel = formatLevel(noColor)
-	output.FormatMessage = func(i interface{}) string {
-		return fmt.Sprintf("| %-60s ", i)
-	}
-	log.Logger = log.Output(output)
-
-	// Default level for this example is info, unless debug flag is present
-	zerolog.SetGlobalLevel(zerolog.InfoLevel)
-	if debug {
-		zerolog.SetGlobalLevel(zerolog.DebugLevel)
-	}
+	setupLogOutput()
 }
 
 func main() {
@@ -76,57 +62,26 @@ func main() {
 			log.Error().Int("obj#", i).Err(err).Msg("error writing object")
 		}
 
-		fmt.Fprint(w, string(f.Bytes()))
+		formatted := formatObject(f.Bytes())
+
+		fmt.Fprint(w, string(formatted))
 		fmt.Fprintln(w)
 	}
-
 }
 
-func formatLevel(noColor bool) zerolog.Formatter {
-	return func(i interface{}) string {
-		var l string
-		if ll, ok := i.(string); ok {
-			switch ll {
-			case "debug":
-				l = colorize("Debug", colorYellow, noColor)
-			case "info":
-				l = colorize("Info", colorGreen, noColor)
-			case "warn":
-				l = colorize("Warn", colorRed, noColor)
-			case "error":
-				l = colorize(colorize("Error", colorRed, noColor), colorBold, noColor)
-			case "fatal":
-				l = colorize(colorize("Fatal", colorRed, noColor), colorBold, noColor)
-			case "panic":
-				l = colorize(colorize("Panic", colorRed, noColor), colorBold, noColor)
-			default:
-				l = colorize("???", colorBold, noColor)
-			}
-		} else {
-			l = strings.ToUpper(fmt.Sprintf("%s", i))[0:3]
+func formatObject(in []byte) ([]byte) {
+	var result []byte
+	var err error
+
+	if tf12format {
+		result = hclwrite.Format(in)
+	} else {
+		result, err = printer.Format(in)
+		if err != nil {
+			log.Error().Err(err).Msg("could not format object")
+			return in
 		}
-		return l
 	}
+
+	return result
 }
-
-// colorize returns the string s wrapped in ANSI code c, unless disabled is true.
-func colorize(s interface{}, c int, disabled bool) string {
-	if disabled {
-		return fmt.Sprintf("%s", s)
-	}
-	return fmt.Sprintf("\x1b[%dm%v\x1b[0m", c, s)
-}
-
-const (
-	colorBlack = iota + 30
-	colorRed
-	colorGreen
-	colorYellow
-	colorBlue
-	colorMagenta
-	colorCyan
-	colorWhite
-
-	colorBold     = 1
-	colorDarkGray = 90
-)
